Issue #377 - use the VMUSB drivers in place rather than as copies
Import support for caenchain directly from vmusb/daqconfig with
appropriate #ifdefery.
